Moved DataGridComponent to a seperate folder.
