+++ tools/gulptasks/dashboards/api-docs.json
@@ -27,7 +27,8 @@
         "../../../ts/Dashboards/Layout/Layout.ts",
         "../../../ts/Dashboards/Layout/Row.ts",
         "../../../ts/Dashboards/PluginHandler.ts",
-        "../../../ts/Dashboards/Plugins/DataGridComponent.ts",
+        "../../../ts/Dashboards/Plugins/DataGridComponent/DataGridComponent.ts",
+        "../../../ts/Dashboards/Plugins/DataGridComponent/DataGridComponentOptions.ts",
         "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponent.ts",
         "../../../ts/Dashboards/Plugins/HighchartsComponent/HighchartsComponentOptions.ts",
         "../../../ts/Dashboards/Plugins/KPIComponent.ts",
