+++ ts/Dashboards/Plugins/DataGridComponent/DataGridComponent.ts
@@ -19,25 +19,25 @@
  *
  * */
 
-import type Cell from '../Layout/Cell';
-import type DataGrid from '../../DataGrid/DataGrid';
-import type DataTable from '../../Data/DataTable';
-import type BaseDataGridOptions from '../../DataGrid/DataGridOptions';
-import type { ColumnOptions } from '../../DataGrid/DataGridOptions';
-import type MathModifierOptions from '../../Data/Modifiers/MathModifierOptions';
-import type SidebarPopup from '../EditMode/SidebarPopup';
-import type Sync from '../Components/Sync/Sync';
+import type Cell from '../../Layout/Cell';
+import type DataGrid from '../../../DataGrid/DataGrid';
+import type DataTable from '../../../Data/DataTable';
+import type BaseDataGridOptions from '../../../DataGrid/DataGridOptions';
+import type { ColumnOptions } from '../../../DataGrid/DataGridOptions';
+import type DataConnectorType from '../../../Data/Connectors/DataConnectorType';
+import type MathModifierOptions from '../../../Data/Modifiers/MathModifierOptions';
+import type Options from './DataGridComponentOptions';
+import type SidebarPopup from '../../EditMode/SidebarPopup';
 
-import Component from '../Components/Component.js';
-import DataConnector from '../../Data/Connectors/DataConnector.js';
-import DataConverter from '../../Data/Converters/DataConverter.js';
+import Component from '../../Components/Component.js';
+import DataConnector from '../../../Data/Connectors/DataConnector.js';
+import DataConverter from '../../../Data/Converters/DataConverter.js';
 import DataGridSyncHandlers from './DataGridSyncHandlers.js';
-import U from '../../Core/Utilities.js';
-import DataConnectorType from '../../Data/Connectors/DataConnectorType';
+import DataGridComponentDefaults from './DataGridComponentDefaults.js';
+import U from '../../../Core/Utilities.js';
 const {
     diffObjects,
-    merge,
-    uniqueKey
+    merge
 } = U;
 
 /* *
@@ -67,18 +67,7 @@ class DataGridComponent extends Component {
     /** @private */
     public static defaultOptions = merge(
         Component.defaultOptions,
-        {
-            dataGridClassName: 'dataGrid-container',
-            dataGridID: 'dataGrid-' + uniqueKey(),
-            dataGridOptions: {},
-            editableOptions: [{
-                name: 'connectorName',
-                propertyPath: ['connector', 'id'],
-                type: 'select'
-            }],
-            syncHandlers: DataGridSyncHandlers,
-            onUpdate: DataGridComponent.onUpdate
-        }
+        DataGridComponentDefaults
     );
 
     /* *
@@ -86,66 +75,6 @@ class DataGridComponent extends Component {
      *  Static Functions
      *
      * */
-
-    /**
-     * Default update function, if data grid has changed. This functionality can
-     * be replaced with the {@link DataGridComponent.DataGridOptions#onUpdate}
-     * option.
-     *
-     * @private
-     *
-     * @param e
-     * Related keyboard event of the change.
-     *
-     * @param connector
-     * Relate store of the change.
-     */
-    public static onUpdate(
-        e: KeyboardEvent,
-        connector: Component.ConnectorTypes
-    ): void {
-        const inputElement = e.target as HTMLInputElement;
-        if (inputElement) {
-            const parentRow = inputElement
-                .closest('.highcharts-datagrid-row');
-            const cell = inputElement.closest('.highcharts-datagrid-cell');
-
-            const converter = new DataConverter();
-
-            if (
-                parentRow &&
-                parentRow instanceof HTMLElement &&
-                cell &&
-                cell instanceof HTMLElement
-            ) {
-                const dataTableRowIndex = parentRow
-                    .dataset.rowIndex;
-                const { columnName } = cell.dataset;
-
-                if (
-                    dataTableRowIndex !== void 0 &&
-                    columnName !== void 0
-                ) {
-                    const table = connector.table;
-
-                    if (table) {
-                        let valueToSet = converter
-                            .asGuessedType(inputElement.value);
-
-                        if (valueToSet instanceof Date) {
-                            valueToSet = valueToSet.toString();
-                        }
-
-                        table.setCell(
-                            columnName,
-                            parseInt(dataTableRowIndex, 10),
-                            valueToSet
-                        );
-                    }
-                }
-            }
-        }
-    }
 
     /** @private */
     public static fromJSON(
@@ -157,7 +86,7 @@ class DataGridComponent extends Component {
 
         const component = new DataGridComponent(
             cell,
-            merge<DataGridComponent.Options>(options as any, {
+            merge<Options>(options as any, {
                 dataGridOptions,
                 syncHandlers: DataGridComponent.syncHandlers
             })
@@ -179,12 +108,16 @@ class DataGridComponent extends Component {
 
     /** @private */
     public dataGrid?: DataGrid;
+
     /** @private */
     public dataGridOptions: Partial<BaseDataGridOptions>;
+
     /** @private */
-    public options: DataGridComponent.Options;
+    public options: Options;
+
     /** @private */
     public sync: Component['sync'];
+
     /** @private */
     private connectorListeners: Array<Function>;
 
@@ -194,16 +127,13 @@ class DataGridComponent extends Component {
      *
      * */
 
-    constructor(
-        cell: Cell,
-        options: Partial<DataGridComponent.Options>
-    ) {
+    constructor(cell: Cell, options: Partial<Options>) {
         options = merge(DataGridComponent.defaultOptions, options);
 
         super(cell, options);
 
         this.connectorListeners = [];
-        this.options = options as DataGridComponent.Options;
+        this.options = options as Options;
         this.type = 'DataGrid';
 
         if (this.options.dataGridClassName) {
@@ -375,18 +305,13 @@ class DataGridComponent extends Component {
     }
 
     /** @private */
-    public resize(
-        width?: number|null,
-        height?: number|null
-    ): void {
+    public resize(width?: number|null, height?: number|null): void {
         if (this.dataGrid) {
             super.resize(width, height);
         }
     }
 
-    public async update(
-        options: Partial<DataGridComponent.Options>
-    ): Promise<void> {
+    public async update(options: Partial<Options>): Promise<void> {
         if (options.connector?.id !== this.connectorId) {
             const connectorListeners = this.connectorListeners;
             for (let i = 0, iEnd = connectorListeners.length; i < iEnd; ++i) {
@@ -478,10 +403,10 @@ class DataGridComponent extends Component {
         }
     }
 
-    public getOptionsOnDrop(sidebar: SidebarPopup): Partial<DataGridComponent.Options> {
+    public getOptionsOnDrop(sidebar: SidebarPopup): Partial<Options> {
         const connectorsIds =
             sidebar.editMode.board.dataPool.getConnectorIds();
-        let options: Partial<DataGridComponent.Options> = {
+        let options: Partial<Options> = {
             cell: '',
             type: 'DataGrid'
         };
@@ -523,7 +448,7 @@ class DataGridComponent extends Component {
      * @internal
      *
      */
-    public getOptions(): Partial<DataGridComponent.Options> {
+    public getOptions(): Partial<Options> {
         return {
             ...diffObjects(this.options, DataGridComponent.defaultOptions),
             type: 'DataGrid'
@@ -566,130 +491,6 @@ namespace DataGridComponent {
         json: ClassJSON;
     }
     >;
-
-    /**
-     * Options to control the DataGrid component.
-     */
-    export interface Options extends Component.Options {
-
-        /**
-         * The style class to add to the rendered data grid container.
-         */
-        dataGridClassName?: string;
-
-        /**
-         * The identifier for the rendered data grid container.
-         */
-        dataGridID?: string;
-
-        /**
-         * Callback to use when a change in the data grid occurs.
-         */
-        onUpdate: typeof DataGridComponent.onUpdate
-
-        type: 'DataGrid';
-        /**
-         * Generic options to adjust behavior and styling of the rendered data
-         * grid.
-         */
-        dataGridOptions?: BaseDataGridOptions;
-
-        /**
-         * The set of options like `dataGridClassName` and `dataGridID`.
-         */
-        chartClassName?: string;
-
-        /**
-         * The id that is applied to the chart's container.
-         */
-        chartID?: string;
-
-        /**
-         * Defines which elements should be synced.
-         * ```
-         * Example:
-         * {
-         *     highlight: true
-         * }
-         * ```
-         * Try it:
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/demo/sync-extremes/ | Extremes Sync }
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/component-options/sync-highlight/ | Highlight Sync }
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/component-options/sync-visibility/ | Visibility Sync }
-         *
-         */
-        sync?: SyncOptions;
-
-        /** @private */
-        tableAxisMap?: Record<string, string | null>;
-
-        /**
-         * If the `visibleColumns` option is not provided, the data grid will
-         * calculate and include each column from the data connector.
-         * When declared, the data grid will only include the columns that are
-         * listed.
-         *
-         * Alternatively, the column visibility can be controlled by the
-         * `dataGridOptions.columns` option.
-         * ```
-         * Example
-         * visibleColumns: ['Food', 'Vitamin A']
-         * ```
-         */
-        visibleColumns?: Array<string>;
-
-    }
-
-    /**
-     * Sync options available for the DataGrid component.
-     *
-     * Example:
-     * ```
-     * {
-     *     highlight: true
-     * }
-     * ```
-     */
-    export interface SyncOptions extends Sync.RawOptionsRecord {
-        /**
-         * Extremes sync is available for Highcharts, KPI, DataGrid and
-         * Navigator components. Sets a common range of displayed data. For the
-         * KPI Component sets the last value.
-         *
-         * Try it:
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/demo/sync-extremes/ | Extremes Sync }
-         *
-         * @default false
-         */
-        extremes?: boolean|Sync.OptionsEntry;
-        /**
-         * Highlight sync is available for Highcharts and DataGrid components.
-         * It allows to highlight hovered corresponding rows in the table and
-         * chart points.
-         *
-         * Try it:
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/component-options/sync-highlight/ | Highlight Sync }
-         *
-         * @default false
-         */
-        highlight?: boolean|Sync.HighlightSyncOptions;
-        /**
-         * Visibility sync is available for Highcharts and DataGrid components.
-         * Synchronizes the visibility of data from a hidden/shown series.
-         *
-         * Try it:
-         *
-         * {@link https://jsfiddle.net/gh/get/library/pure/highcharts/highcharts/tree/master/samples/dashboards/component-options/sync-visibility/ | Visibility Sync }
-         *
-         * @default false
-         */
-        visibility?: boolean|Sync.OptionsEntry;
-    }
 
     /** @private */
     export interface ComponentJSONOptions
